+++ fs_attachment/models/ir_model_fields.py
@@ -5,6 +5,7 @@ from odoo import fields, models
 
 
 class IrModelFields(models.Model):
+
     _inherit = "ir.model.fields"
 
     storage_id = fields.Many2one(